ai_services/main.py
"""
NeuroBridge+ AI Services
Serviços de Inteligência Artificial para personalização e inclusão
"""

from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional, Dict, Any
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(
    title="NeuroBridge+ AI Services",
    description="Serviços de IA para personalização de experiências inclusivas",
    version="1.0.0",
)

# CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=[os.getenv("FRONTEND_URL", "http://localhost:3000")],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Modelos de Dados
class AccessibilityAssessment(BaseModel):
    """Avaliação de acessibilidade do usuário"""
    disability_types: List[str]
    communication_preferences: List[str]
    learning_style: str
    assistive_technologies: List[str]
    answers: Dict[str, Any]

class ProfileResponse(BaseModel):
    """Resposta do perfil gerado"""
    profile_id: str
    recommendations: List[str]
    accessibility_score: float
    personalization_tips: List[str]

class CourseRecommendationRequest(BaseModel):
    """Requisição de recomendação de cursos"""
    user_id: str
    profile_id: str
    interests: Optional[List[str]] = []
    limit: int = 10

class JobRecommendationRequest(BaseModel):
    """Requisição de recomendação de vagas"""
    user_id: str
    profile_id: str
    skills: List[str]
    location_preference: Optional[str] = None
    limit: int = 10

class ChatMessage(BaseModel):
    """Mensagem do chat"""
    user_id: str
    message: str
    format: str = "text"  # text, audio, libras
    context: Optional[Dict[str, Any]] = {}

class ChatResponse(BaseModel):
    """Resposta do chat"""
    message: str
    format: str
    translated: Optional[Dict[str, str]] = {}
    suggestions: Optional[List[str]] = []

# Rotas

@app.get("/")
async def root():
    """Rota raiz"""
    return {
        "service": "NeuroBridge+ AI Services",
        "version": "1.0.0",
        "status": "operational",
        "endpoints": {
            "profile": "/api/profile",
            "recommendations": "/api/recommendations",
            "chat": "/api/chat",
            "accessibility": "/api/accessibility",
        }
    }

@app.get("/health")
async def health_check():
    """Health check"""
    return {
        "status": "healthy",
        "services": {
            "profile_engine": "operational",
            "recommender": "operational",
            "chatbot": "operational",
            "nlp_processor": "operational",
        }
    }

# Profile Engine
@app.post("/api/profile/generate", response_model=ProfileResponse)
async def generate_profile(assessment: AccessibilityAssessment):
    """
    Gera um perfil de acessibilidade personalizado baseado na avaliação
    """
    # TODO: Implementar lógica real do profile engine
    return ProfileResponse(
        profile_id="profile_123",
        recommendations=[
            "Ativar modo de alto contraste",
            "Usar fonte tamanho grande",
            "Habilitar legendas automáticas",
        ],
        accessibility_score=0.85,
        personalization_tips=[
            "Seus cursos serão adaptados para aprendizado visual",
            "Recomendamos usar audiodescrição em vídeos",
        ]
    )

@app.get("/api/profile/{profile_id}")
async def get_profile(profile_id: str):
    """
    Recupera um perfil de acessibilidade existente
    """
    # TODO: Implementar recuperação do banco de dados
    return {
        "profile_id": profile_id,
        "status": "active",
        "last_updated": "2024-01-20T10:00:00Z",
    }

# Recommender System
@app.post("/api/recommendations/courses")
async def recommend_courses(request: CourseRecommendationRequest):
    """
    Recomenda cursos personalizados baseados no perfil do usuário
    """
    # TODO: Implementar sistema de recomendação real
    return {
        "user_id": request.user_id,
        "recommendations": [
            {
                "course_id": "course_1",
                "title": "Introdução à Programação Python",
                "score": 0.95,
                "reasons": [
                    "Compatível com seu estilo de aprendizado visual",
                    "Possui transcrição e legendas",
                    "Texto simplificado disponível",
                ],
            },
            {
                "course_id": "course_2",
                "title": "Design Inclusivo e Acessibilidade",
                "score": 0.92,
                "reasons": [
                    "Alinhado com seus interesses",
                    "Formato adaptado para neurodivergentes",
                ],
            },
        ],
        "total": 2,
    }

@app.post("/api/recommendations/jobs")
async def recommend_jobs(request: JobRecommendationRequest):
    """
    Recomenda vagas personalizadas baseadas no perfil e habilidades
    """
    # TODO: Implementar sistema de matching de vagas
    return {
        "user_id": request.user_id,
        "recommendations": [
            {
                "job_id": "job_1",
                "title": "Desenvolvedor Python Júnior",
                "company": "Tech Inclusiva LTDA",
                "score": 0.88,
                "reasons": [
                    "Ambiente de trabalho acessível",
                    "Horário flexível",
                    "Suporte para neurodivergentes",
                ],
                "accessibility_match": 0.95,
            },
        ],
        "total": 1,
    }

# Chatbot
@app.post("/api/chat/message", response_model=ChatResponse)
async def chat_message(message: ChatMessage):
    """
    Processa mensagem do chat e retorna resposta adaptada
    """
    # TODO: Implementar chatbot com LLM
    return ChatResponse(
        message="Olá! Como posso ajudá-lo hoje?",
        format=message.format,
        translated={
            "simplified": "Oi! Posso ajudar você?",
            "audio_url": "https://example.com/audio.mp3",
        },
        suggestions=[
            "Quero encontrar cursos",
            "Procuro emprego",
            "Ajuda com acessibilidade",
        ]
    )

@app.post("/api/chat/translate")
async def translate_message(message: ChatMessage):
    """
    Traduz mensagem para diferentes formatos (texto simples, áudio, Libras)
    """
    # TODO: Implementar tradução multimodal
    return {
        "original": message.message,
        "translations": {
            "simplified_text": "Versão simplificada da mensagem",
            "audio_url": "https://example.com/audio.mp3",
            "libras_video_url": "https://example.com/libras.mp4",
        }
    }

# Accessibility Audit
@app.post("/api/accessibility/audit")
async def audit_accessibility(url: str):
    """
    Audita a acessibilidade de um site
    """
    # TODO: Implementar auditoria de acessibilidade
    return {
        "url": url,
        "score": 75,
        "wcag_level": "AA",
        "issues": [
            {
                "severity": "high",
                "type": "contrast",
                "message": "Contraste insuficiente entre texto e fundo",
                "element": "button.primary",
            },
            {
                "severity": "medium",
                "type": "alt_text",
                "message": "Imagem sem texto alternativo",
                "element": "img#hero",
            },
        ],
        "recommendations": [
            "Aumentar contraste para pelo menos 4.5:1",
            "Adicionar atributo alt em todas as imagens",
            "Implementar navegação por teclado",
        ],
    }

# NLP Processor
@app.post("/api/nlp/simplify")
async def simplify_text(text: str, level: str = "basic"):
    """
    Simplifica texto para facilitar compreensão
    """
    # TODO: Implementar simplificação de texto com NLP
    return {
        "original": text,
        "simplified": "Versão simplificada do texto",
        "level": level,
        "readability_score": 85,
    }

@app.post("/api/nlp/extract-skills")
async def extract_skills(text: str):
    """
    Extrai habilidades de um texto (CV, descrição de vaga, etc.)
    """
    # TODO: Implementar extração de habilidades com NLP
    return {
        "skills": [
            {"name": "Python", "confidence": 0.95},
            {"name": "JavaScript", "confidence": 0.88},
            {"name": "React", "confidence": 0.82},
        ],
        "categories": ["Programação", "Frontend", "Backend"],
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=8000,
        reload=True,
        log_level="info",
    )

