ai_services/requirements.txt
# Framework Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Validação de Dados
pydantic==2.5.3
pydantic-settings==2.1.0

# Banco de Dados
motor==3.3.2
pymongo==4.6.1
redis==5.0.1

# Machine Learning e IA
openai==1.7.2
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
numpy==1.26.3
pandas==2.1.4

# Processamento de Linguagem Natural
nltk==3.8.1
spacy==3.7.2

# Processamento de Voz
openai-whisper==20231117
gtts==2.5.0

# Banco Vetorial
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Utilitários
python-dotenv==1.0.0
httpx==0.26.0
aiohttp==3.9.1

# Logging e Monitoramento
loguru==0.7.2

# Testes
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.26.0

